Eliminate the Work.abstract property and move the Document.content property to Work.content instead.  Update affected views, and re-base some classes that were deriving from Document just to get Document.content to now be derived from Work instead.
The Document class now has a "pass" body with no additional properties.  It
is debatable whether Work should become Document, since I am not sure that
we plan to be able to generically display, say, a Quiz or a Question (which
are Works) on the home page (which can only specifically display a Document,
or now, the basic Work entity...).

Patch by: Todd Larsen
Review by: to-be-reviewed