Add explicit access_types from the url
This does two things, it reduces the amount of repeated data (e.g.,
the checkAccess used to repeat the access_type that was already
specified in the url). At the same time, it allows for example the
'create' page to push its access_type to 'edit'.

Patch by: Sverre Rabbelier