Make it possible to invite another Host as Host
We do this by asking Django to resolve the current url for us, then
reshufle that url and ask Django (again) to resolve that new url.
From that resolved url we know the appropriate view, and as such we
know which rights dictioanry to pass to access.checkAccess.

Patch by: Sverre Rabbelier