Add an export() view, and implement it as text/text for Document.
For every Model except Document, the public() view is displayed for any
attempts to access the export() view.

Currently, the permissions for export() are the same as for public().  This
seems reasonable for Document, since anyone could extract the raw HTML from
the page source anyway.  The permissions should probably be different for
other types of exports, such as vCard or iCard exports of profiles, CSV
exports of lists, etc.

Patch by: Todd Larsen
Review by: to-be-reviewed