Fix get args construction when there are multiple lists on the page
It is now possible to go back and forward through the liast, and
specify the limit (both offset and limit can be done per list).
The JS driving the list boxes is buggy, if visiting an url like:
http://localhost:8080/notification/list?limit_0=10
And then change the limit in the second checkbox, it directs to:
http://localhost:8080/notification/list?limit_1=25
Whereas it should redirect to:
http://localhost:8080/notification/list?limit_0=10&limit_1=25
The logic _does_ work properly when the limit of the changed list
is already present in the url.

Patch by: Sverre Rabbelier