Remove the release.py module and just use the field in app.yaml.template.
This is a DRY fix, to have the release /branches/ and /tags/ use a release
version that is compatible with the app.yaml file "version:" field.

Also, the steps for creating a release were simplified (with several steps
removed), and the release procedure moved to app/README, since
app/soc/release.py is now deleted.

Patch by:  Todd Larsen
Review by: to-be-reviewed